chat_message/serializers.py
from rest_framework import serializers


from chat_message.models import ChatMessage


class ChatMessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = ChatMessage
        fields = '__all__'
