requirements.txt
dj-database-url==0.4.2
Django
djangorestframework==3.9.1
gunicorn==19.7.1
pytz==2017.3
whitenoise==3.3.1
psycopg2-binary==2.8.4
